feat(app): do not move referenced issues via board
This board removes automatic related issue movement, as it leads
to extra noise on the board and polution of the IN DEV / REVIEW columns.

When moving issues on the board we previously moved all issues
that are marked as CLOSED BY the issue with the issue.

Users can trace the dependency based on issue links and update
the related issues manually.